wip refactoring for omni tiny models
