+++ tests/model_tests/utils.py
@@ -1,80 +1,41 @@
 """
-Common definitions for controlling what tests run where.
+Utilities for resolving real models to their tiny model equivalents.
 """
 
-from collections.abc import Callable
-from enum import StrEnum, auto
-from typing import NamedTuple, TypeAlias
+import logging
 
-from pytest import MarkDecorator
-
-from vllm_omni.diffusion.data import DiffusionParallelConfig
+from tests.model_tests.config_types import DiffusionAccs
+from tests.model_tests.diffusion.model_settings import DIFFUSION_TEST_SETTINGS
+from vllm_omni.diffusion.data import DiffusionParallelConfig, resolve_model_class_name
 from vllm_omni.entrypoints.omni import Omni
 
-# All builder funcs take no params and return a path
-TinyDiffusionBuilder: TypeAlias = Callable[[], str]
+logger = logging.getLogger(__name__)
 
 
-class DiffusionAccs(StrEnum):
-    """Supported acceleration types / test settings for Diffusion Models."""
+def resolve_tiny_model_path(model: str) -> str:
+    """Given a real model name/path, resolve it to a tiny model path.
 
-    HSDP = auto()
-    TEA_CACHE = auto()
-    CACHE_DIT = auto()
-    SEQUENCE_PARALLEL = auto()
-    CFG_PARALLEL = auto()
-    TENSOR_PARALLEL = auto()
-    CPU_OFFLOAD = auto()
-    LAYERWISE_OFFLOAD = auto()
-    VAE_PATCH_PARALLEL = auto()
+    Raises ValueError if the pipeline class cannot be determined (invalid
+    model). Returns the original model path if no tiny builder exists yet."""
+    pipeline_class = resolve_model_class_name(model)
+    if pipeline_class is None:
+        raise ValueError(
+            f"Cannot resolve pipeline class for model: {model}. The model path may be invalid or its config unreadable."
+        )
+
+    test_opts = DIFFUSION_TEST_SETTINGS.get(pipeline_class)
+    if test_opts is None:
+        logger.warning(
+            "No tiny model builder for pipeline %s (model: %s). Using original model.",
+            pipeline_class,
+            model,
+        )
+        return model
+
+    return test_opts.builder()
 
 
-class DiffusionTasks(StrEnum):
-    """Supported tasks for Diffusion Models."""
-
-    TEXT_TO_IMAGE = auto()
-    IMAGE_TO_IMAGE = auto()
-    TEXT_TO_VIDEO = auto()
-    # Text to audio, etc should be added here as needed
-
-
-class DiffusionModelTestOpts(NamedTuple):
-    """Configuration for one Diffusion model's tests."""
-
-    # HF model name for real-weight tests (advanced_model / full_model level).
-    # For now, whether we use the real weights vs tiny weights in the common tests
-    # depends on the run level.
-    model: str
-
-    # Creates a tiny model for the given architecture. We should always use tiny
-    # model weights for tests that do not require us to check the model quality.
-    builder: TinyDiffusionBuilder
-
-    # Actual tasks which controls the tests actually run
-    supported_tasks: list[DiffusionTasks]
-
-    # Additional acceleration groups to run beyond the base case (no acceleration).
-    # The base case is always run for every model in the test settings. None means
-    # we only run the base case.
-    extra_test_groups: list[list[DiffusionAccs]] | None = None
-
-    # Pytest Marks for this model. This may be useful for selecting which models
-    # we want to run where, similar to the way vLLM's multimodal tests mark some
-    # as core models to always run in the CI.
-    # Example: https://github.com/vllm-project/vllm/blob/v0.23.0/tests/models/multimodal/generation/test_common.py#L131
-    marks: list[MarkDecorator] | None = None
-
-    # When True (default), online tests only run the base case (no accelerations).
-    # When False, online tests run all test_groups, same as offline. This should be
-    # True unless there is a good reason for it not to be, because the execution of
-    # the acceleration should be the same on both codepaths, and CLI parsing etc should
-    # be tested by tests adding the acceleration, and not per model.
-    online_base_only: bool = True
-
-    # Additional checks to run for the base case.
-    check_multi_output: bool = True  # Runs multiple generations in one request
-    check_determinism: bool = True  # Runs 2 generations with the same seed and check determinism
-
+### helpers for building Diffusion Models
 
 ### Mappings & utils for building offline Omni() instances given a list of enabled accelerations
 ACC_OMNI_KWARGS = {
